Extensions partial revert
