+++ src/typechat/StringEx.cs
@@ -1,5 +1,9 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
+using System;
+using System.Collections.Generic;
+using System.Text;
+
 namespace Microsoft.TypeChat;
 
 internal static class StringEx
